mcp-browser-co-gnome/docker/browser/requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
pydantic>=2.0.0
pyyaml>=6.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
pynput>=1.7.6
python-xlib>=0.33
pillow>=10.0.0
